[Feat] 기능 명세서 조회
이슈 삭제 로직 일부 수정
Service 레이어에서 Entity 반환

+++ be/src/main/java/com/sharep/be/modules/issue/service/IssueServiceImpl.java
@@ -5,17 +5,12 @@ import com.sharep.be.modules.api.repository.ApiRepository;
 import com.sharep.be.modules.assignee.Assignee;
 import com.sharep.be.modules.assignee.State;
 import com.sharep.be.modules.assignee.repository.AssigneeRepository;
-import com.sharep.be.modules.exception.AssigneeNotFoundException;
 import com.sharep.be.modules.exception.IssueNotFoundException;
 import com.sharep.be.modules.exception.MemberNotFoundException;
 import com.sharep.be.modules.exception.ProjectNotFoundException;
-import com.sharep.be.modules.exception.UnauthorizedException;
 import com.sharep.be.modules.issue.Issue;
 import com.sharep.be.modules.issue.IssueRequest.IssueCreate;
 import com.sharep.be.modules.issue.IssueRequest.IssueUpdate;
-import com.sharep.be.modules.issue.IssueResponse;
-import com.sharep.be.modules.issue.IssueResponse.IssueCreated;
-import com.sharep.be.modules.issue.IssueResponse.PrivateIssueResponse;
 import com.sharep.be.modules.issue.repository.IssueRepository;
 import com.sharep.be.modules.issue.type.IssueType;
 import com.sharep.be.modules.member.Member;
@@ -23,10 +18,11 @@ import com.sharep.be.modules.member.MemberRepository;
 import com.sharep.be.modules.project.Project;
 import com.sharep.be.modules.project.ProjectRepository;
 import jakarta.transaction.Transactional;
-import java.util.Collections;
 import java.util.List;
+import java.util.Set;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
+import org.springframework.util.Assert;
 
 @Service
 @Transactional
@@ -40,11 +36,11 @@ public class IssueServiceImpl implements IssueService {
     private final AssigneeRepository assigneeRepository;
 
     @Override
-    public IssueCreated createIssue(Long projectId, Long accountId, IssueCreate issueCreate) {
+    public Issue createIssue(Long projectId, Long accountId, IssueCreate issueCreate) {
         Project project = projectRepository.findById(projectId)
                 .orElseThrow(ProjectNotFoundException::new);
 
-        Issue issue = issueRepository.save(Issue.from(issueCreate, project));
+        Issue issue = issueRepository.save(issueCreate.toEntityWith(project));
         apiRepository.save(Api.builder().issue(issue).build());
 
         IssueType type = issueCreate.type();
@@ -52,58 +48,62 @@ public class IssueServiceImpl implements IssueService {
             Member member = memberRepository.findByAccountIdAndProjectId(accountId, projectId)
                     .orElseThrow(MemberNotFoundException::new);
 
-            assigneeRepository.save(
-                    Assignee.builder().issue(issue).member(member).state(State.YET).build());
-
+            assigneeRepository.save(Assignee.builder()
+                    .issue(issue)
+                    .member(member)
+                    .state(State.YET)
+                    .build());
         }
 
-        return issue.toCreated();
-    }
-
-
-    @Override
-    public List<IssueResponse> getIssues(Long projectId) {
-        return issueRepository.findIssuesByProjectId(projectId).orElse(Collections.emptyList())
-                .stream().map(Issue::toResponse).toList();
+        return issue;
     }
 
     @Override
-    public IssueResponse getIssue(Long id) {
-        return issueRepository.findById(id).orElseThrow(IssueNotFoundException::new).toResponse();
+    public Issue getIssue(Long id) {
+
+        return issueRepository.findById(id).orElseThrow(IssueNotFoundException::new);
     }
 
     @Override
-    public List<PrivateIssueResponse> getPrivateIssues(Long projectId, Long accountId) {
+    public List<Issue> getPrivateIssues(Long projectId, Long accountId) {
         Member member = memberRepository.findByAccountIdAndProjectId(accountId, projectId)
                 .orElseThrow(MemberNotFoundException::new);
 
-        return issueRepository.findIssuesByMemberIdAndIssueType(member.getId(), IssueType.PRIVATE)
-                .stream().map(Issue::toPrivateIssueResponse).toList();
+        return issueRepository.findIssuesByMemberIdAndIssueType(member.getId(), IssueType.PRIVATE);
+    }
+
+    @Override
+    public List<Issue> getFeatureIssues(Long projectId) {
+
+        return issueRepository.findIssuesByProjectIdAndIssueType(projectId, IssueType.FEATURE);
     }
 
     @Override
     public void updateIssue(Long id, Long accountId, Long projectId, IssueUpdate issueUpdate) {
         Issue issue = issueRepository.findById(id).orElseThrow(IssueNotFoundException::new);
+        assertPrivateIssueModifyAuthorization(accountId, projectId, issue);
+        issueRepository.save(issueUpdate.toEntityWith(issue));
+    }
+
+    public void deleteIssue(Long id, Long accountId, Long projectId) {
+        Issue issue = issueRepository.findById(id).orElseThrow(IssueNotFoundException::new);
+        assertPrivateIssueModifyAuthorization(accountId, projectId, issue);
+        issueRepository.deleteById(id);
+    }
+
+    private void assertPrivateIssueModifyAuthorization(Long accountId, Long projectId,
+            Issue issue) {
         IssueType type = issue.getType();
         if (type.equals(IssueType.PRIVATE)) {
             Long memberId = memberRepository.findByAccountIdAndProjectId(accountId, projectId)
                     .orElseThrow(MemberNotFoundException::new).getId();
 
-            Assignee assignee = issue.getAssignees().stream().findFirst()
-                    .orElseThrow(AssigneeNotFoundException::new);
+            Set<Assignee> assignees = issue.getAssignees();
+            Assert.isTrue(assignees.size() == 1, "");
 
-            if (!assignee.getMember().getId().equals(memberId)) {
-                throw new UnauthorizedException();
-            }
+            Assignee assignee = assignees.stream().findFirst().get();
+            Assert.isTrue(assignee.getMember().getId().equals(memberId), "");
         }
-
-        issueRepository.save(issue.from(issueUpdate));
-    }
-
-    @Override
-    public void deleteIssue(Long id) {
-        // TODO: 권한 설정
-        issueRepository.deleteById(id);
     }
 
 }
